refactor: replace accordion picker with drill-through picker
