+++ frontend/src/lib/components/BreadcrumbSegment.svelte
@@ -1,24 +1,27 @@
 <!--
 @component
 One clickable segment of the editor breadcrumb. The segment opens a popover
-containing a `WorkspaceItemPicker`. Each instance owns its own `isOpen`
+containing a `WorkspaceItemDrillPicker`. Each instance owns its own `isOpen`
 state — switching popovers relies on melt-ui's `closeOnOtherPopoverOpen` to
 close siblings.
 -->
 <script lang="ts">
 	import { ChevronRight } from 'lucide-svelte'
 	import Popover from '$lib/components/meltComponents/Popover.svelte'
-	import WorkspaceItemPicker, {
-		type WorkspaceItem
-	} from '$lib/components/WorkspaceItemPicker.svelte'
+	import WorkspaceItemDrillPicker, {
+		type Scope
+	} from '$lib/components/WorkspaceItemDrillPicker.svelte'
+	import type { WorkspaceItem } from '$lib/components/workspacePicker'
 
 	interface Props {
 		label: string
 		withChevron?: boolean
 		extraClass?: string
 		disabled?: boolean
-		initialOpen: string[]
-		initialHighlight: string
+		/** Where the picker lands when this segment is clicked. */
+		initialScope?: Scope
+		/** Composite key of the row to highlight (e.g. `dir:flow:f/demo`). */
+		initialHighlight?: string
 		currentItem?: WorkspaceItem & { savedPath?: string }
 		onPick: (item: WorkspaceItem) => void
 	}
@@ -28,7 +31,7 @@ close siblings.
 		withChevron = false,
 		extraClass = '',
 		disabled = false,
-		initialOpen,
+		initialScope,
 		initialHighlight,
 		currentItem,
 		onPick
@@ -61,6 +64,6 @@ close siblings.
 			>{:else}{label}{/if}
 	{/snippet}
 	{#snippet content()}
-		<WorkspaceItemPicker {initialOpen} {initialHighlight} {currentItem} onPick={handlePick} />
+		<WorkspaceItemDrillPicker {initialScope} {initialHighlight} {currentItem} onPick={handlePick} />
 	{/snippet}
 </Popover>
