--- frontend/src/lib/components/WorkspaceItemPicker.svelte
@@ -1,587 +1,0 @@
-<script lang="ts" module>
-	import type { Flow, ListableApp, Script } from '$lib/gen'
-
-	export type WorkspaceItemKind = 'flow' | 'script' | 'app'
-
-	/** Shape emitted by the picker's `onPick` callback. The `raw_app` flag is only
-	 * meaningful when `kind === 'app'` and discriminates low-code (`false`) from
-	 * raw apps (`true`); see `editPathFor` for routing. */
-	export type WorkspaceItem = {
-		path: string
-		summary: string
-		kind: WorkspaceItemKind
-		raw_app?: boolean
-	}
-
-	type Kind = WorkspaceItemKind
-	type Item = WorkspaceItem
-
-	/** Resolve the editor URL for an item picked from the workspace picker.
-	 * Apps split into low-code (`/apps/edit`) and raw (`/apps_raw/edit`). */
-	export function editPathFor(item: WorkspaceItem): string {
-		if (item.kind === 'flow') return `/flows/edit/${item.path}`
-		if (item.kind === 'script') return `/scripts/edit/${item.path}`
-		return item.raw_app ? `/apps_raw/edit/${item.path}` : `/apps/edit/${item.path}`
-	}
-
-	type WorkspaceCache = {
-		flow?: Item[]
-		script?: Item[]
-		app?: Item[]
-	}
-
-	const cache = new Map<string, WorkspaceCache>()
-	const inflight = new Map<string, Promise<Item[]>>()
-
-	function cacheKey(workspace: string, kind: Kind) {
-		return `${workspace}:${kind}`
-	}
-
-	async function loadKind(workspace: string, kind: Kind): Promise<Item[]> {
-		const existing = cache.get(workspace)?.[kind]
-		if (existing) return existing
-		const key = cacheKey(workspace, kind)
-		const flying = inflight.get(key)
-		if (flying) return flying
-
-		const promise = (async () => {
-			const { ScriptService, FlowService, AppService } = await import('$lib/gen')
-			let items: Item[]
-			if (kind === 'flow') {
-				const flows = await FlowService.listFlows({
-					workspace,
-					includeDraftOnly: true,
-					withoutDescription: true
-				})
-				items = flows.map((f: Flow) => ({
-					path: f.path,
-					summary: f.summary ?? '',
-					kind: 'flow' as const
-				}))
-			} else if (kind === 'script') {
-				const scripts = await ScriptService.listScripts({
-					workspace,
-					includeDraftOnly: true,
-					withoutDescription: true
-				})
-				items = scripts.map((s: Script) => ({
-					path: s.path,
-					summary: s.summary ?? '',
-					kind: 'script' as const
-				}))
-			} else {
-				const apps = await AppService.listApps({
-					workspace,
-					includeDraftOnly: true
-				})
-				items = apps.map((a: ListableApp) => ({
-					path: a.path,
-					summary: a.summary ?? '',
-					kind: 'app' as const,
-					raw_app: a.raw_app ?? false
-				}))
-			}
-			const bucket = cache.get(workspace) ?? {}
-			bucket[kind] = items
-			cache.set(workspace, bucket)
-			return items
-		})()
-		inflight.set(key, promise)
-		try {
-			return await promise
-		} finally {
-			inflight.delete(key)
-		}
-	}
-</script>
-
-<script lang="ts">
-	import { workspaceStore } from '$lib/stores'
-	import { ChevronRight, Folder, Loader2, User } from 'lucide-svelte'
-	import TextInput from '$lib/components/text_input/TextInput.svelte'
-	import RowIcon from '$lib/components/common/table/RowIcon.svelte'
-	import SearchItems from '$lib/components/SearchItems.svelte'
-	import type uFuzzy from '@leeoniya/ufuzzy'
-
-	type Kind_ = 'flow' | 'script' | 'app'
-
-	interface Props {
-		onPick: (item: WorkspaceItem) => void
-		kinds?: Kind_[]
-		/** Composite-keyed nodes to open on mount. e.g. `['kind:flow', 'scope:flow:f/demo']`. */
-		initialOpen?: string[]
-		/** Composite key to highlight on mount. Preserved even if not yet visible (e.g. while
-		 * the kind is still loading). */
-		initialHighlight?: string
-		/** Inject a virtual entry for the currently-edited item so it appears at its
-		 * live (possibly draft-renamed) path. If `savedPath` differs from `path`,
-		 * the saved entry is removed from the list to avoid duplicates. */
-		currentItem?: WorkspaceItem & { savedPath?: string }
-	}
-
-	let {
-		onPick,
-		kinds = ['flow', 'script', 'app'],
-		initialOpen = [],
-		initialHighlight,
-		currentItem
-	}: Props = $props()
-
-	let searchInput: TextInput | undefined = $state()
-	let pickerRoot: HTMLElement | undefined = $state()
-
-	/** Focus + select the search input. Callable from outside (e.g. via a popover's openFocus). */
-	export function focus() {
-		searchInput?.focus()
-		searchInput?.select()
-	}
-
-	// Fallback for switching between sibling breadcrumb popovers (closeOnOtherPopoverOpen):
-	// melt-ui's `openFocus` runs once during the close→open transition and the new
-	// picker may not be mounted yet when it tries to query the search input. Retry
-	// after the transition has settled.
-	$effect(() => {
-		setTimeout(() => focus(), 50)
-	})
-
-	const KIND_LABEL: Record<Kind_, string> = {
-		flow: 'Flows',
-		script: 'Scripts',
-		app: 'Apps'
-	}
-
-	const kindKey = (k: Kind_) => `kind:${k}`
-	const dirKey = (k: Kind_, fullPath: string) => `dir:${k}:${fullPath}`
-	const leafKey = (it: Item) => `leaf:${it.kind}:${it.path}`
-
-	let filter = $state('')
-	let openItems = $state<Set<string>>(new Set(initialOpen))
-
-	// Seed from the module-level cache so kinds already fetched in this session
-	// render their scopes/leaves on the first frame (no async microtask gap, no
-	// flash of "kind label only" while opening).
-	let loaded = $state<Partial<Record<Kind_, Item[]>>>(
-		(() => {
-			const ws = $workspaceStore && cache.get($workspaceStore)
-			if (!ws) return {}
-			const out: Partial<Record<Kind_, Item[]>> = {}
-			for (const k of kinds) if (ws[k]) out[k] = ws[k]
-			return out
-		})()
-	)
-	let loadingKind = $state<Partial<Record<Kind_, boolean>>>({})
-
-	async function ensureLoaded(kind: Kind_) {
-		if (!$workspaceStore) return
-		if (loaded[kind]) return
-		loadingKind[kind] = true
-		try {
-			const items = await loadKind($workspaceStore, kind)
-			loaded[kind] = items
-		} finally {
-			loadingKind[kind] = false
-		}
-	}
-
-	// Lazy-load each kind on first expansion
-	$effect(() => {
-		for (const id of openItems) {
-			if (id.startsWith('kind:')) {
-				const k = id.slice(5) as Kind_
-				if (kinds.includes(k)) ensureLoaded(k)
-			}
-		}
-	})
-
-	// While searching, ensure every kind in scope is loaded so the search is comprehensive.
-	$effect(() => {
-		if (filter.trim() !== '') {
-			for (const k of kinds) {
-				ensureLoaded(k)
-			}
-		}
-	})
-
-	type DirNode = {
-		/** Full path prefix, e.g. "f/demo" or "f/demo/sub". */
-		fullPath: string
-		/** Last segment of the path. For top-level dirs (scopes), the full scope
-		 * (e.g. "f/demo"); for nested dirs, just the leaf segment. */
-		name: string
-		/** True for a top-level scope dir; controls icon (User vs Folder). */
-		isScope: boolean
-		children: DirNode[]
-		leaves: Item[]
-	}
-
-	/** Build a directory tree per kind from a flat item list. Each item's slug
-	 * (path parts after the scope owner) is split on `/`; intermediate parts
-	 * become nested dir nodes, the last part is the leaf. */
-	function buildTrees(): Record<Kind_, DirNode[]> {
-		const out: Record<Kind_, DirNode[]> = { flow: [], script: [], app: [] }
-		for (const k of kinds) {
-			if (!loaded[k] && !(currentItem && currentItem.kind === k)) continue
-			const list = itemsForKind(k)
-			const scopeRoots = new Map<string, DirNode>()
-			for (const it of list) {
-				const parts = it.path.split('/')
-				if (parts.length < 3) continue
-				const scope = parts.slice(0, 2).join('/')
-				let node = scopeRoots.get(scope)
-				if (!node) {
-					node = { fullPath: scope, name: scope, isScope: true, children: [], leaves: [] }
-					scopeRoots.set(scope, node)
-				}
-				const slug = parts.slice(2)
-				let cur = node
-				for (let i = 0; i < slug.length - 1; i++) {
-					const seg = slug[i]
-					const fullPath = cur.fullPath + '/' + seg
-					let next = cur.children.find((c) => c.name === seg)
-					if (!next) {
-						next = { fullPath, name: seg, isScope: false, children: [], leaves: [] }
-						cur.children.push(next)
-					}
-					cur = next
-				}
-				cur.leaves.push(it)
-			}
-			// Sort scopes (folders first, then alpha) and recurse for children/leaves.
-			const scopes = Array.from(scopeRoots.values()).sort((a, b) => {
-				const af = a.fullPath.startsWith('f/') ? 0 : 1
-				const bf = b.fullPath.startsWith('f/') ? 0 : 1
-				if (af !== bf) return af - bf
-				return a.fullPath.localeCompare(b.fullPath)
-			})
-			const sortNode = (n: DirNode) => {
-				n.children.sort((a, b) => a.name.localeCompare(b.name))
-				n.leaves.sort((a, b) => a.path.localeCompare(b.path))
-				n.children.forEach(sortNode)
-			}
-			scopes.forEach(sortNode)
-			out[k] = scopes
-		}
-		return out
-	}
-
-	type DisplayItem = Item & { marked?: string }
-	type SearchInput = Item & { _key: string }
-
-	/** Items for a kind, with the current draft-renamed item swapped in:
-	 * remove the saved entry (if it differs) and append the live one. Lets the
-	 * breadcrumb's draft path always have a corresponding scope/leaf to expand. */
-	function itemsForKind(k: Kind_): Item[] {
-		const base = loaded[k] ?? []
-		if (!currentItem || currentItem.kind !== k) return base
-		const drafted =
-			currentItem.savedPath && currentItem.savedPath !== currentItem.path
-				? base.filter((it) => it.path !== currentItem.savedPath)
-				: base
-		return drafted.some((it) => it.path === currentItem.path)
-			? drafted
-			: [
-					...drafted,
-					{
-						path: currentItem.path,
-						summary: currentItem.summary,
-						kind: k,
-						raw_app: currentItem.raw_app
-					}
-				]
-	}
-
-	let allItems = $derived<SearchInput[]>(
-		kinds.flatMap((k) => itemsForKind(k).map((it) => ({ ...it, _key: `${k}:${it.path}` })))
-	)
-
-	let searchedItems: DisplayItem[] | undefined = $state(undefined)
-
-	const searchOpts: uFuzzy.Options = {}
-
-	let dirTreeByKind = $derived.by(() => buildTrees())
-
-	let isSearching = $derived(filter.trim() !== '')
-
-	let searchResultsByKind = $derived.by(() => {
-		const out: Record<Kind_, DisplayItem[]> = { flow: [], script: [], app: [] }
-		if (!searchedItems) return out
-		for (const it of searchedItems) {
-			out[it.kind].push(it)
-		}
-		return out
-	})
-
-	function isCurrent(it: Item): boolean {
-		return !!currentItem && currentItem.kind === it.kind && currentItem.path === it.path
-	}
-
-	function pick(it: Item) {
-		// No-op when the user clicks their currently-edited item — they're already there.
-		if (isCurrent(it)) return
-		onPick({ path: it.path, summary: it.summary, kind: it.kind, raw_app: it.raw_app })
-	}
-
-	type NavNode =
-		| { type: 'kind'; key: string; kind: Kind_ }
-		| { type: 'dir'; key: string; kind: Kind_; node: DirNode }
-		| { type: 'leaf'; key: string; item: Item }
-
-	let navNodes: NavNode[] = $derived.by(() => {
-		if (isSearching) {
-			return kinds.flatMap((k) =>
-				searchResultsByKind[k].map((it): NavNode => ({ type: 'leaf', key: leafKey(it), item: it }))
-			)
-		}
-		const nodes: NavNode[] = []
-		const walk = (k: Kind_, dir: DirNode) => {
-			const dKey = dirKey(k, dir.fullPath)
-			nodes.push({ type: 'dir', key: dKey, kind: k, node: dir })
-			if (!openItems.has(dKey)) return
-			for (const child of dir.children) walk(k, child)
-			for (const it of dir.leaves) nodes.push({ type: 'leaf', key: leafKey(it), item: it })
-		}
-		for (const k of kinds) {
-			const kKey = kindKey(k)
-			nodes.push({ type: 'kind', key: kKey, kind: k })
-			if (!openItems.has(kKey)) continue
-			if (!loaded[k] && !(currentItem && currentItem.kind === k)) continue
-			for (const scope of dirTreeByKind[k]) walk(k, scope)
-		}
-		return nodes
-	})
-
-	let highlightedKey = $state<string | undefined>(initialHighlight)
-
-	// Default to the first node when nothing is highlighted yet. In search mode
-	// we also reset to the first node when the current highlight is no longer
-	// visible (the search result set changed). In tree mode we preserve a non-
-	// visible highlight — it might become visible after a lazy load completes.
-	$effect(() => {
-		if (navNodes.length === 0) return
-		if (isSearching) {
-			if (!highlightedKey || !navNodes.some((n) => n.key === highlightedKey)) {
-				highlightedKey = navNodes[0].key
-			}
-		} else if (!highlightedKey) {
-			highlightedKey = navNodes[0].key
-		}
-	})
-
-	// Once the caller-supplied highlight becomes visible, scroll it into view.
-	$effect(() => {
-		if (highlightedKey && navNodes.some((n) => n.key === highlightedKey)) {
-			queueMicrotask(scrollHighlightIntoView)
-		}
-	})
-
-	function scrollHighlightIntoView() {
-		if (!pickerRoot || !highlightedKey) return
-		const el = pickerRoot.querySelector<HTMLElement>(
-			`[data-nav-key="${CSS.escape(highlightedKey)}"]`
-		)
-		el?.scrollIntoView({ block: 'nearest' })
-	}
-
-	function moveHighlight(delta: 1 | -1) {
-		if (navNodes.length === 0) return
-		const cur = navNodes.findIndex((n) => n.key === highlightedKey)
-		const next = cur < 0 ? 0 : (cur + delta + navNodes.length) % navNodes.length
-		highlightedKey = navNodes[next].key
-		queueMicrotask(scrollHighlightIntoView)
-	}
-
-	function toggleOpen(key: string) {
-		const next = new Set(openItems)
-		if (next.has(key)) next.delete(key)
-		else next.add(key)
-		openItems = next
-	}
-
-	function activate(key: string | undefined) {
-		if (!key) return
-		const node = navNodes.find((n) => n.key === key)
-		if (!node) return
-		if (node.type === 'leaf') {
-			pick(node.item)
-		} else {
-			toggleOpen(node.key)
-		}
-	}
-
-	function handleSearchKeydown(e: KeyboardEvent) {
-		if (e.key === 'ArrowDown') {
-			e.preventDefault()
-			e.stopPropagation()
-			moveHighlight(1)
-		} else if (e.key === 'ArrowUp') {
-			e.preventDefault()
-			e.stopPropagation()
-			moveHighlight(-1)
-		} else if (e.key === 'Enter') {
-			e.preventDefault()
-			e.stopPropagation()
-			activate(highlightedKey)
-		}
-	}
-</script>
-
-<SearchItems
-	{filter}
-	items={isSearching ? allItems : []}
-	bind:filteredItems={searchedItems}
-	f={(x: SearchInput) => (x.summary ? `${x.summary} (${x.path})` : x.path)}
-	opts={searchOpts}
-/>
-
-{#snippet leafRow(it: Item, secondary: string, baseClass: string, paddingLeftRem: number = 0)}
-	{@const key = leafKey(it)}
-	{@const isHl = key === highlightedKey}
-	{@const isCur = isCurrent(it)}
-	<button
-		type="button"
-		data-nav-key={key}
-		aria-current={isCur ? 'page' : undefined}
-		class="w-full text-left flex items-center gap-2 transition-colors {baseClass} {isHl
-			? 'bg-surface-hover'
-			: ''} {isCur ? 'cursor-default text-emphasis font-medium' : ''}"
-		style={paddingLeftRem ? `padding-left: ${paddingLeftRem}rem` : ''}
-		onclick={() => pick(it)}
-		onmouseenter={() => (highlightedKey = key)}
-	>
-		<RowIcon kind={it.kind} size={12} />
-		<div class="min-w-0 flex-1">
-			{#if it.summary}
-				<div class="text-xs text-primary truncate">{it.summary}</div>
-				<div class="text-2xs text-secondary font-normal font-mono truncate">{secondary}</div>
-			{:else}
-				<div class="text-xs text-primary font-mono truncate">{secondary}</div>
-			{/if}
-		</div>
-	</button>
-{/snippet}
-
-{#snippet dirSubtree(k: Kind_, dir: DirNode, depth: number)}
-	{@const dKey = dirKey(k, dir.fullPath)}
-	{@const isOpen = openItems.has(dKey)}
-	{@const isHl = dKey === highlightedKey}
-	<button
-		type="button"
-		data-nav-key={dKey}
-		aria-expanded={isOpen}
-		class="flex items-center gap-1.5 w-full text-left pr-3 py-1.5 text-xs font-medium font-mono text-emphasis transition-colors {isHl
-			? 'bg-surface-hover'
-			: ''}"
-		style="padding-left: {0.75 + depth * 1.25}rem"
-		onclick={() => toggleOpen(dKey)}
-		onmouseenter={() => (highlightedKey = dKey)}
-	>
-		<ChevronRight
-			size={10}
-			class="transition-transform shrink-0 text-secondary {isOpen ? 'rotate-90' : ''}"
-		/>
-		{#if dir.isScope && dir.fullPath.startsWith('u/')}
-			<User size={12} class="shrink-0 text-tertiary" />
-		{:else}
-			<Folder size={12} class="shrink-0 text-tertiary" />
-		{/if}
-		<span class="truncate">{dir.name}</span>
-	</button>
-	{#if isOpen}
-		{#each dir.children as child (child.fullPath)}
-			{@render dirSubtree(k, child, depth + 1)}
-		{/each}
-		{#each dir.leaves as it (it.path)}
-			{@render leafRow(
-				it,
-				it.path.slice(dir.fullPath.length + 1),
-				'pr-3 py-1.5',
-				0.75 + (depth + 1) * 1.25
-			)}
-		{/each}
-	{/if}
-{/snippet}
-
-<!-- svelte-ignore a11y_no_static_element_interactions -->
-<div
-	bind:this={pickerRoot}
-	class="flex flex-col w-[420px] max-h-[60vh]"
-	onkeydown={handleSearchKeydown}
->
-	<div class="px-3 py-2 border-b border-gray-200 dark:border-gray-700">
-		<TextInput
-			bind:this={searchInput}
-			bind:value={filter}
-			size="sm"
-			inputProps={{
-				placeholder: 'Search by name or summary...',
-				'data-workspace-picker-search': ''
-			}}
-		/>
-	</div>
-
-	<div class="flex-1 overflow-y-auto">
-		{#if isSearching}
-			{@const total = (searchedItems ?? []).length}
-			{#if !searchedItems}
-				<div class="px-3 py-2 text-xs text-tertiary flex items-center gap-2">
-					<Loader2 size={14} class="animate-spin" /> Searching…
-				</div>
-			{:else if total === 0}
-				<div class="px-3 py-2 text-xs text-tertiary">No matches</div>
-			{:else}
-				{#each kinds as k (k)}
-					{@const results = searchResultsByKind[k]}
-					{#if results.length > 0}
-						<div class="px-3 pt-3 pb-1 text-2xs uppercase tracking-wide text-tertiary font-medium">
-							{KIND_LABEL[k]}
-						</div>
-						<ul class="pb-1">
-							{#each results as it (leafKey(it))}
-								<li>{@render leafRow(it, it.path, 'px-3 py-1.5')}</li>
-							{/each}
-						</ul>
-					{/if}
-				{/each}
-			{/if}
-		{:else}
-			<div class="flex flex-col py-1">
-				{#each kinds as k (k)}
-					{@const kKey = kindKey(k)}
-					{@const kindOpen = openItems.has(kKey)}
-					{@const kindHl = kKey === highlightedKey}
-					<button
-						type="button"
-						data-nav-key={kKey}
-						aria-expanded={kindOpen}
-						class="flex items-center gap-1.5 w-full text-left px-3 py-1.5 text-xs font-medium font-mono text-emphasis transition-colors {kindHl
-							? 'bg-surface-hover'
-							: ''}"
-						onclick={() => toggleOpen(kKey)}
-						onmouseenter={() => (highlightedKey = kKey)}
-					>
-						<ChevronRight
-							size={10}
-							class="transition-transform shrink-0 text-secondary {kindOpen ? 'rotate-90' : ''}"
-						/>
-						<RowIcon kind={k} size={12} />
-						<span class="flex-1">{KIND_LABEL[k]}</span>
-						{#if loadingKind[k]}
-							<Loader2 size={12} class="animate-spin text-tertiary" />
-						{/if}
-					</button>
-					{#if kindOpen}
-						{#if !loaded[k] && !loadingKind[k] && !(currentItem && currentItem.kind === k)}
-							<div class="px-8 py-1 text-2xs text-tertiary">Empty</div>
-						{:else if loaded[k]?.length === 0 && !(currentItem && currentItem.kind === k)}
-							<div class="px-8 py-1 text-2xs text-tertiary">No {KIND_LABEL[k].toLowerCase()}</div>
-						{:else}
-							{#each dirTreeByKind[k] as scope (scope.fullPath)}
-								{@render dirSubtree(k, scope, 1)}
-							{/each}
-						{/if}
-					{/if}
-				{/each}
-			</div>
-		{/if}
-	</div>
-</div>
